Update tests for compiler changes (#72)
Since Xcode 14 / Swift 5.7, some file paths in index units that were
previously stored relatively, are now stored absolutely. Because of this
our remaps have to change to support the new structure. Along with this,
also as of Xcode 14, `swiftc` and `clang` support remapping the paths in
indexes using `-file-prefix-map foo=bar` and `-ffile-prefix-map=foo=bar`
respectively. This can be used to remove the absolute paths entirely
from the produced index data, which then also requires a new set of
remaps to convert back to absolute for Xcode. I've mostly updated the
tests to use this remapping behavior, which everyone should use if
possible, but I left one case with updated paths to show the new
absolute path behavior.

Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>