Add `-file-prefix-map` option
The `-file-prefix-map` option allows setting the `PathRemapper` mappings the same way that `clang` does. We now also use the `PathRemapper` in `getUnitPathForOutputFile` the same as as `IndexUnitWriter`.

This change is needed to allow the use of both the `swift.use_global_index_store` and `swift.file_prefix_map` rules_swift features.

Signed-off-by: Brentley Jones <github@brentleyjones.com>