Revert "Absolutize paths after remapping (#91)"
Absolutizing these paths resulted in failures with symlinks downstream.
Resolving those symlinks seems to work but in the case you're trying to
remap from 1 location to another not for Xcode, which rules_swift does
if you use the `swift.use_global_index_store` you actually need the
paths to still be relative in the units.

This reverts commit 431d4e5d1e75baa7316a66919160dc682ee93863.

Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>